colab/requirements.txt
firebase-admin==6.2.0
openai==1.3.0
pandas
matplotlib
numpy
google-generativeai
